PR changes: * move save_results() to utils
