+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] ; 2025-06-25
+
+### Changed
+
+- Added wrapper function `run_evaluate()` to improve workflow
+- Return results as df
+- `save_results()` moved outside Evaluator class
+
 ## [0.3.0] ; 2025-06-25
 
 ### Added
