+++ mmm_eval/cli/evaluate.py
@@ -1,36 +1,15 @@
 import logging
-from datetime import datetime
-from pathlib import Path
 
 import click
-import pandas as pd
 
 from mmm_eval.adapters import ADAPTER_REGISTRY
 from mmm_eval.configs import get_config
 from mmm_eval.core import run_evaluation
 from mmm_eval.core.validation_tests_models import ValidationTestNames
 from mmm_eval.data.loaders import DataLoader
+from mmm_eval.utils import save_results
 
 logger = logging.getLogger(__name__)
-
-
-def save_results(results: pd.DataFrame, framework: str, output_path: str) -> None:
-    """Save the results to a CSV file.
-
-    Args:
-        results: The dataframe of results to save.
-        framework: The name of the framework that was evaluated.
-        output_path: The path to save the results to.
-
-    """
-    output_path_obj = Path(output_path)
-    output_path_obj.mkdir(parents=True, exist_ok=True)
-
-    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-    filename = f"mmm_eval_{framework}_{timestamp}.csv"
-
-    results.to_csv(output_path_obj / filename)
-    logger.info(f"Saved results to {output_path_obj / filename}")
 
 
 @click.command()
@@ -55,6 +34,7 @@ def save_results(results: pd.DataFrame, framework: str, output_path: str) -> Non
 @click.option(
     "--config-path",
     type=str,
+    required=True,
     help="Path to framework-specific JSON config file",
 )
 @click.option(
